+++ examples/snippets/sms/groups/replace/snippet.py
@@ -1,0 +1,30 @@
+"""
+Sinch Python Snippet
+
+This snippet is available at https://github.com/sinch/sinch-sdk-python/tree/main/examples/snippets
+"""
+
+import os
+
+from dotenv import load_dotenv
+
+from sinch import SinchClient
+
+load_dotenv()
+
+sinch_client = SinchClient(
+    project_id=os.environ.get("SINCH_PROJECT_ID") or "MY_PROJECT_ID",
+    key_id=os.environ.get("SINCH_KEY_ID") or "MY_KEY_ID",
+    key_secret=os.environ.get("SINCH_KEY_SECRET") or "MY_KEY_SECRET",
+    sms_region=os.environ.get("SINCH_SMS_REGION") or "MY_SMS_REGION"
+)
+
+# The ID of the group to replace
+group_id = "GROUP_ID"
+
+response = sinch_client.sms.groups.replace(
+    group_id=group_id,
+    members=["+1234567890", "+1987654321"],
+)
+
+print(f"Group replaced:\n{response}")