+++ sinch/domains/sms/sms.py
@@ -3,6 +3,7 @@ from sinch.domains.sms.api.v1 import (
     DeliveryReports,
 )
 from sinch.domains.sms.api.v1.inbounds_apis import Inbounds
+from sinch.domains.sms.api.v1.groups_apis import Groups
 from sinch.domains.sms.sinch_events.v1.sms_sinch_event import SmsSinchEvent
 
 
@@ -18,6 +19,7 @@ class SMS:
         self.batches = Batches(self._sinch)
         self.delivery_reports = DeliveryReports(self._sinch)
         self.inbound_messages = Inbounds(self._sinch)
+        self.groups = Groups(self._sinch)
 
     def sinch_events(self, sinch_event_secret: str) -> SmsSinchEvent:
         """